requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
pydub>=0.25.1
speechrecognition>=3.10.0
pyaudio>=0.2.14
pygame>=2.5.0
flask>=3.0.0
flask-cors>=4.0.0
protobuf>=3.20.0
python-dotenv>=1.0.0

